Refactor: #248 전체 상태 조회에서 단일 상태 조회 로직 분리
