+++ src/components/modal/project/ModalProjectForm.tsx
@@ -10,13 +10,13 @@ import SearchUserInput from '@components/common/SearchUserInput';
 import UserRoleSelectBox from '@components/common/UserRoleSelectBox';
 import DescriptionTextarea from '@components/common/DescriptionTextarea';
 import DuplicationCheckInput from '@components/common/DuplicationCheckInput';
-import { getProjectNameList } from '@utils/extractNameList';
 import useAxios from '@hooks/useAxios';
 import useToast from '@hooks/useToast';
 import { useReadProjects } from '@hooks/query/useProjectQuery';
 import { findUserByTeam } from '@services/teamService';
 import { useParams } from 'react-router-dom';
 import type { SubmitHandler } from 'react-hook-form';
+import { getProjectNameList } from '@/utils/extractDataList';
 import type { TeamSearchCallback } from '@/types/SearchCallbackType';
 import type { ProjectRoleName } from '@/types/RoleType';
 import type { ProjectCoworker, ProjectForm } from '@/types/ProjectType';
